WIP: forking
FIXME: Currently this starts Two Of Them, I have no idea why.

Right now we use a library for forking that appears to be unmaintained
upstream, while we also forward our stdin/stdout to a set of files named
with our username. This is pretty annoying, because it means one nvim
instance will clobber the logging of another if it starts up, along with
the fact it would be much nicer for us to simply fork and inherit the
stdout/stderr of our original terminal much like other editors (such as the
original GTK+ support in vanilla vim) so we can still print messages to the
user.

So, let's use a different library and fix that.