Adjust cursor focus logic for GTK4
We have pretty much everything else working, so let's get focus fixed. We
currently sort of handle this with a FocusEventController, but now that we
have both the toplevel window being "active" in addition to the
NvimViewport being focused we need to make the rendering of the cursor
depend on being both active and focused.