popup_menu: Don't store reference to row
I realized we don't actually have any reason for doing this - we pretty
much only need the information in the row when we're setting up the
CompletionListRow. We'll keep around the row state at the very least though
so we don't have to pass it every time we bind a CompletionListRow, however
we might in the future want to consider getting rid of that.