React to pmenu(_sel) and cursor highlight updates
The main purpose of this is to make it so that instead of updating the CSS
of the popup menu every single time we get a new list of completion items,
we instead only update it upon receiving updates to Pmenu or PmenuSel.
While I haven't actually profiled this, it's pretty easy to guess that it's
likely not ideal to be updating the CSS styles of a ListView widget with a
very large number of children needlessly.

Also since the code for this is right nearby: we also take this opportunity
to trigger cursor redraws when the Cursor highlight group changes.