Remove nvim-background class on nvim error
Right now we leave the nvim-background class on the application window,
which means that if we ever display an error we'll mistakenly do it with a
totally transparent window background. So, be careful to remove this class
in situations where we won't be relying on the viewport to set the
background color.