Handle popup_menu events more in line with actual flushes
I mistakenly made it so that we were flushing popupmenu events whenever we
finished receiving a batch of events, when we really only want to be
flushing all of the events we've received since the most recent flush.

So, let's fix that. We also move pending_popupmenu back into the shell
state since it turns out it can be there - and kind of makes more sense
since that's where we store the pending redraw state as well.