shell: Don't allow main gtk::Paned to become focusable
Apparently we forgot to make sure this couldn't get keyboard focus, which
causes an obscure bug where after switching windows - focus is suddenly on
the wrong widget, which results in keyboard interactions with nvim-gtk
behaving improperly. I'm not sure what the exact trigger for this is, but
I've hit it every now and then during development and only just finally
managed to catch it with the gtk inspector.