Add CI workflow
Let's finally start using Github's CI! This makes it so we run the
following checks using a Fedora 37 container:

  * cargo test
  * cargo fmt --check
  * git status (to make sure Cargo.lock doesn't go out of sync)

The reason I chose Fedora is because it makes it easier to get the latest
stable GTK+, and also I like Fedora. There shouldn't be much that would
break that's distro-specific anyway though, so distro choice shouldn't
matter too much here. We still need to do something for OSX though, and
Windows.